mcp-natural-language-to-data-endpoints/requirements.txt
# MCP Server Core
mcp>=0.9.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# SQL Databases
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
cx-Oracle>=8.3.0
pyodbc>=5.0.0
snowflake-sqlalchemy>=1.5.0
databricks-sql-connector>=3.0.0

# NoSQL Databases
pymongo>=4.6.0
cassandra-driver>=3.28.0
redis>=5.0.0
boto3>=1.34.0  # For DynamoDB

# Graph Databases
neo4j>=5.15.0
python-arango>=7.8.0
SPARQLWrapper>=2.0.0
gremlinpython>=3.7.0

# GraphQL
httpx>=0.26.0

# Utilities
asyncio>=3.4.3
